+++ django_test_urls/__init__.py
@@ -1,14 +1,14 @@
 #!/usr/bin/env python3
 # -*- coding: UTF-8 -*-
 
-""" A package used to test the mapping of URLs to views and arguments.
+""" A package used to test the mapping of URLs to views and parameters.
 
-:copyright: (c) 2022 by Alan Verresen.
+:copyright: (c) 2022 by Alan Verresen
 :license: MIT, see LICENSE for more details.
 """
 
 from .resolves_to import resolves_to
-from .resolves_to import resolves_to_args
+from .resolves_to import resolves_to_arguments
 from .resolves_to import resolves_to_view
 from .resolves_to import resolves_to_404
 
@@ -16,7 +16,7 @@ from .resolves_to import resolves_to_404
 __all__ = (
     'resolves_to',
     'resolves_to_404',
-    'resolves_to_args',
+    'resolves_to_arguments',
     'resolves_to_view',
 )