+++ tests/app_settings.py
@@ -2,7 +2,8 @@
 # -*- coding: UTF-8 -*-
 
 from django.urls import include
-from django.urls import re_path
+from django.urls import path
+
 
 ROOT_URLCONF = __name__
-urlpatterns = [re_path(r"^articles/", include('tests.app_urls'))]
+urlpatterns = [path("", include('tests.app_urls'))]